Ciência-da-Computação/PROJETOS/Project-tech-news/sd-020-b-tech-news/tech_news/analyzer/search_engine.py
from tech_news.database import search_news
from datetime import datetime


# Requisito 6
def search_by_title(title):
    news = search_news({"title": {"$regex": title, "$options": "i"}})
    return [(new["title"], new["url"]) for new in news]


# Requisito 7
def search_by_date(date):
    try:
        timestamp = datetime.strptime(date, "%Y-%m-%d").strftime("%d/%m/%Y")
    except ValueError:
        raise ValueError("Data inválida")
    else:
        news = search_news({"timestamp": timestamp})
        return [(new["title"], new["url"]) for new in news]


# Requisito 8
def search_by_tag(tag):
    news = search_news(
        {"tags": {"$elemMatch": {"$regex": tag, "$options": "i"}}}
    )
    return [(new["title"], new["url"]) for new in news]


# Requisito 9
def search_by_category(category):
    news = search_news({"category": {"$regex": category, "$options": "i"}})
    return [(new["title"], new["url"]) for new in news]
